requirements.txt
accelerate==1.0.1
aiohttp==3.10.11
async-timeout==5.0.1
contourpy==1.1.1
cycler==0.12.1
datasets==3.1.0
diffusers==0.23.0
dill==0.3.8
filelock==3.16.1
fonttools==4.57.0
fsspec==2024.9.0
hf-xet==1.1.5
huggingface-hub==0.25.0
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
multiprocess==0.70.16
networkx==3.1
numpy==1.24.4
packaging==25.0
pandas==2.0.3
pillow==10.4.0
propcache==0.2.0
psutil==7.0.0
pyarrow==17.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
sentencepiece==0.2.0
sympy==1.13.3
tokenizers==0.19.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.67.1
transformers==4.44.2
triton==3.0.0
tzdata==2025.2
urllib3==2.2.3
xxhash==3.5.0
yarl==1.15.2
